update: refractor onnx for shared code between nodejs/browser
