+++ packages/gpt2-onnx/src/node/tokenizer.ts
@@ -1,9 +1,3 @@
-import { Tokenizer } from "@llama-node/hf-tokenizer";
-import { Tensor } from "onnxruntime-node";
 import { IsomorphicTokenizer } from "../shared/tokenizer";
 
-export class AutoTokenizer extends IsomorphicTokenizer {
-    constructor() {
-        super(Tokenizer, Tensor);
-    }
-}
+export class AutoTokenizer extends IsomorphicTokenizer {}
