+++ packages/gpt2-onnx/src/shared/tokenizer.ts
@@ -1,18 +1,19 @@
 import type { Tokenizer } from "@llama-node/hf-tokenizer";
 import type { TensorConstructor } from "onnxruntime-node";
+import { loadHfTokenizer, loadOnnx } from "./loader";
 
 export class IsomorphicTokenizer {
     tokenizer?: Tokenizer;
-
-    constructor(
-        private tokenizerConstructor: new (json: string) => Tokenizer,
-        private tensorConstructor: TensorConstructor
-    ) {}
+    Tensor?: TensorConstructor;
 
     async initFromUrl(url: string) {
         const json = await (await fetch(url)).json();
 
-        this.tokenizer = new this.tokenizerConstructor(JSON.stringify(json));
+        const { Tokenizer } = await loadHfTokenizer();
+        const { Tensor } = await loadOnnx();
+
+        this.Tensor = Tensor;
+        this.tokenizer = new Tokenizer(JSON.stringify(json));
     }
 
     free() {
@@ -24,6 +25,10 @@ export class IsomorphicTokenizer {
     }
 
     tokenize(text: string, addSpecialTokens = true) {
+        if (!this.Tensor) {
+            throw new Error("Onnx is not initialized");
+        }
+
         const encoded = this.tokenizer?.encode(text, addSpecialTokens);
 
         const inputIdArray = BigInt64Array.from(
@@ -32,18 +37,17 @@ export class IsomorphicTokenizer {
 
         encoded?.free();
 
-        const input_ids = new this.tensorConstructor("int64", inputIdArray, [
+        const input_ids = new this.Tensor("int64", inputIdArray, [
             1,
             inputIdArray.length,
         ]);
 
         const attentionMaskArray = inputIdArray.map(() => BigInt(1));
 
-        const attention_mask = new this.tensorConstructor(
-            "int64",
-            attentionMaskArray,
-            [1, attentionMaskArray.length]
-        );
+        const attention_mask = new this.Tensor("int64", attentionMaskArray, [
+            1,
+            attentionMaskArray.length,
+        ]);
 
         return { input_ids, attention_mask };
     }
